# src/api/openapi.yaml
openapi: 3.1.1
info:
  title: Nexxus API
  version: 0.0.11
  summary: HTTP surface of a Nexxus deployment.
  description: |
    Nexxus is a config-driven, real-time backend-as-a-service. This document
    describes the **framework** contract: the routes every deployment exposes
    and the envelope every response carries.

    ### Application-defined shapes

    An application's data models are *runtime configuration*, not source code.
    A deployment serving a fitness app and one serving a chat app expose the
    same `/model/*` routes with entirely different payloads. Where the shape is
    the application's to declare, the schema here says so — see `AppModel` and
    `UserDetails`.

    The same applies to `/auth/{strategy}`: which strategies exist is per
    deployment (`api.auth.availableStrategies`) and per application
    (`auth.strategies`).

    ### Multi-tenancy

    One deployment serves many applications. Almost every request carries an
    `nxx-app-id` header naming which one. It is not a secret and can ship in
    client code.

    ### Authentication

    A session is two tokens, issued by `/user/register`, `/auth/{strategy}` and
    `/device/register`:

    - **`token`** — the access token, sent as `Authorization: Bearer`. It always
      names a **device**, and additionally names a **user** when the application
      has authentication enabled. It expires after the application's
      `session.jwtExpiresIn` (one hour by default). There is no device header —
...

...
servers:
...
paths:
...
  /user/register:
    post:
      tags: [ Users ]
      summary: Create an account
      description: |
        Creates a user account and returns a session, so no separate login
        call is needed. Requires an application with authentication and a
        configured `local` strategy.

        Always creates the account's first device and binds the returned token to
        it. There is no `device` parameter: a new account owns none, so there is
        nothing to reuse or name yet. Rename it afterwards with `PUT /device`.

        Reusing a device applies only when authenticating an account that already
        exists — see `POST /auth/{strategy}`.

        Any property beyond the ones listed is stored as a user **detail** and
        must be declared in the application's `auth.userDetailSchema`. An
        undeclared field is a `400`. See `UserDetails`.
      security: []
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ username, password ]
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                userType:
                  type: string
                  description: Must be a user type the application declares.
                  default: default
              additionalProperties: true
              examples:
                - username: ann@example.com
                  password: correct-horse-battery-staple
                  age: 31
      responses:
        '200':
          $ref: '#/components/responses/Session'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/AppNotFound'
        '409':
          description: That username already exists on this application.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: UserAlreadyExistsException
        '503':
          $ref: '#/components/responses/Unavailable'
  /user/me:
    get:
      tags: [ Users ]
      summary: The authenticated user
      description: |
        Returns the user the token names.

        Reads the token, not the stored record. A change made through
        `PUT /user` appears here once the client refreshes its session
        (`POST /auth/refresh`) or signs in again.
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: The user the token names.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: |
            The token names no user. An application without authentication has
            none to return.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: InvalidAuthMethodException
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AppNotFound'
  /user:
    put:
      tags: [ Users ]
      summary: Update the authenticated user
      description: |
        Applies a patch to the account the token names. There is no user
        parameter.

        Refused paths: `userType`, `authProviders`, `devices`, `createdAt`,
        `updatedAt`, and anything under a `details.$…` key (see `UserDetails`).

        Patching `password` hashes the value server-side and, on an account that
        had none, adds `local` to `authProviders`.
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ patch ]
              properties:
                patch:
                  $ref: '#/components/schemas/JsonPatch'
      responses:
        '200':
          description: The patch was accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AppNotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /device/register:
    post:
      tags: [ Devices ]
      summary: Register a new device
      description: |
        Creates a device and returns a token bound to it. Use that token for
        subsequent requests; the caller's previous token still names the
        previous device.

        This route always creates. To *reuse* an existing device, pass its id as
        a hint when authenticating instead.

        Works on an application without authentication, where the device has no
        owner.
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name ]
              properties:
                name:
                  type: string
                  description: Human-readable, shown in the user's device list.
                  examples: [ Ann's Laptop ]
      responses:
        '200':
          description: The device was created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                    description: A new access token, bound to the device just created.
                  refreshToken:
                    type: string
                    description: The new device's refresh token.
                  device:
                    $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AppNotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
...
  /device:
    get:
      tags: [ Devices ]
      summary: The calling device
      description: |
        Returns the device the token names. There is no device parameter.
      parameters:
        - $ref: '#/components/parameters/AppId'
      responses:
        '200':
          description: The device the token names.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          $ref: '#/components/responses/NoDevice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: |
            The device no longer exists. Register a new one.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        const: NotFoundException
    put:
      tags: [ Devices ]
      summary: Rename the calling device
      description: |
        Renames the device the token names. There is no device parameter.
      parameters:
        - $ref: '#/components/parameters/AppId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ name ]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: The device was renamed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/NoDevice'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AppNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The access token. Issued by `/user/register`, `/auth/{strategy}`,
        `/device/register` and `/auth/refresh`. Names a device, and a user where
        the application has authentication.
  parameters:
    AppId:
      name: nxx-app-id
      in: header
      required: true
      description: |
        Which application this request is for. Not a secret; it can ship in
        client code.
      schema:
        type: string
...
  schemas:
    Error:
      type: object
      description: |
        Every failure uses this shape. `error` is a stable identifier to branch
        on; `message` is for humans and may change.
      required: [ error, message ]
      properties:
        error:
          type: string
          description: Identifies the failure.
        message:
          type: string
        stack:
          type: string
          description: Present only when the deployment runs with `NODE_ENV=dev`.
...
    Device:
      type: object
      properties:
        id:
          type: string
        appId:
          type: string
        userId:
          type: string
          description: Absent on an application without authentication.
        name:
          type: string
        type:
          type: string
          enum: [ volatile, persistent, unknown ]
        status:
          type: string
          enum: [ online, offline ]
        lastSeen:
          type: string
          format: date-time
        transport:
          type: [ string, 'null' ]
          description: |
            The transport this device is connected through. `null` when it is
            not connected.
    User:
      type: object
      description: |
        The user as carried by the token. Never includes a password.
      properties:
        id:
          type: string
        username:
          type: string
        userType:
          type: string
        appId:
          type: string
        authProviders:
          type: array
          items:
            type: string
          description: Which strategies this account can authenticate with.
        details:
          $ref: '#/components/schemas/UserDetails'
    UserDetails:
      type: object
      description: |
        A user's profile fields. Two kinds of key live here.

        **Yours.** Whatever the application declares in `auth.userDetailSchema`
        for that user type. Sending a field it does not declare is a `400`.
        These are the only keys this response contains.

        **The auth providers'.** Each auth strategy enabled on the application
        owns a `$auth_<strategy>` namespace on the stored record — `$auth_local`,
        `$auth_google`, and one per third-party strategy the deployment installs
        — holding what that provider reported about the account (Google
        contributes `name` and `id`). The strategy declares those contents, not
        your `auth.userDetailSchema`.

        They are written at login and are **not returned by the API**. To tell
        which providers an account is linked to, read `authProviders` on the
        user.

        Every key beginning with `$` is reserved. Sending one to
        `/user/register`, or patching one through `PUT /user`, is a `400`.
      additionalProperties: true
...
    JsonPatch:
      type: object
      description: |
        Nexxus's own patch shape — `path` and `value` are parallel arrays, so
        one patch can touch several fields. Not RFC 6902.
      required: [ op, path, value ]
      properties:
        op:
          type: string
          enum: [ replace, append, prepend, incr, decr, remove ]
        path:
          type: array
          items:
            type: string
          description: |
            Dot-separated field paths, e.g. `details.age`. Must be declared in
            the model's schema.
        value:
          type: array
          description: |
            One entry per `path`, in the same order. Each entry is typed by the
            field it targets, so the element type is whatever the application's
            schema declares — a string, a number, a date, an object, an array.
          items:
            description: |
              Any JSON value the target field accepts. Dates may be sent as a
              UNIX timestamp in seconds or as a parseable date string; both are
              stored as seconds.
      examples:
        - op: replace
          path: [ details.age, details.nickname ]
          value: [ 31, ann ]
...
  responses:
    Session:
      description: |
        The session's two tokens, the device they are bound to, and the user they
        name.
      content:
        application/json:
          schema:
            type: object
            properties:
              token:
                type: string
                description: The access token.
              refreshToken:
                type: string
                description: Exchanged at `POST /auth/refresh` for the next pair.
              device:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
              user:
                type: object
                properties:
                  id:
                    type: string
                  username:
                    type: string
...
    BadRequest:
      description: The request is malformed or a parameter is invalid.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    enum:
                      - InvalidParametersException
                      - InvalidAuthMethodException
    Unauthorized:
      description: |
        No token was presented on an application that requires one, or the token
        is invalid or expired. An expired access token
        (`UserTokenExpiredException`) is renewed through `POST /auth/refresh`.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    enum:
                      - NoAuthPresentException
                      - UserAuthenticationFailedException
                      - UserTokenExpiredException
...
    AppNotFound:
      description: No application matches the `nxx-app-id` header.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: ApplicationNotFoundException
...
    NoDevice:
      description: |
        This route acts on the calling device, and the token names none.
        Register a device and use the token it returns.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: InvalidParametersException
    ServerError:
      description: |
        An unexpected failure. The message is generic.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: ServerErrorException
    Unavailable:
      description: |
        An upstream service is currently disconnected. Retry.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  error:
                    const: ServiceUnavailableException
security:
  - bearerAuth: []